backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
workos==5.42.0
PyJWT[crypto]>=2.0.0
httpx==0.28.1
beautifulsoup4==4.13.4
lxml==5.3.1
